Patch window.Touch on "desktop" locally for now
We'll need to update the application eventually but that's a bit of
work to use some of the new idioms and updated components from montage
~v0.13.4

This at least gets the app working on Chrome 29 where window.Touch is
truthy even with no touch input device. This is a short term fix.